+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.Reservation;
-import roomescape.ReservationDto;
+import roomescape.data.vo.Reservation;
+import roomescape.data.dto.ReservationDto;
 import roomescape.repository.ReservationRepository;
 
 @RestController
