+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 
 import java.util.List;
-import roomescape.Reservation;
+import roomescape.data.vo.Reservation;
 
 public interface ReservationRepository {
     long add(final Reservation reservation);
